Added a zopeinstance part to test this with a plain Zope 2.13.
Not sure if that is going to work:
You would need to replace the standard acl_users folder in the Zope root with a PluggableAuthService object.
Technically no problem, and that is what Plone does on install,
but it might give problems for existing sites with existing users.
Something to check later.