Removed minimum required version from plone.testing.
Buildout fails on 4.3 because we require 5.0.0 which is not in the version pins.
We could fix that by pinning plone.testing to one version for all Plone versions,
but that gives a problem with some plone.app.testing versions.
So we would need to the the same there.
And the definite improvement in 5.0.0 seem unneeded to me when we just test a single package.