Fixed test failures due to different Referer handling in zope.testbrowser.
The zope.testbrowser version in Plone 5.2.1 keeps the Referer when calling browser.open for a second time.
That is useful, but not correct, as https://github.com/zopefoundation/zope.testbrowser/issues/87 shows.
Basically: when you manually type a url in a real browser, this browser does not pass a Referer either.
The same should be true in zope.testbrowser when you call browser.open.
So we set the Referer manually in the cases where we need it.